#3 file cleanup and refactors
