+++ Backend/routes/api.js
@@ -1,6 +1,6 @@
 const express = require('express');
 let router = express.Router();
 
-router.use('/user', require("../controllers/users"));
+router.use('/user', require("../controllers/userController"));
 
 module.exports = router;
